requirements.txt
streamlit
sqlalchemy
uvicorn
psycopg2-binary
torch
tensorflow 
matplotlib
torchvision
fastapi
python-multipart 
pillow
